#16 Start working on displaying camera view
