+++ SmartCarMQTTApp/app/build.gradle
@@ -41,4 +41,6 @@ dependencies {
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
+
+    implementation 'pl.droidsonroids.gif:android-gif-drawable:1.2.19'
 }
