+++ SmartCarMQTTApp/app/src/main/java/com/example/smartcarmqttapp/HomeActivity.java
@@ -19,10 +19,7 @@ import org.eclipse.paho.client.mqttv3.MqttMessage;
 public class HomeActivity extends AppCompatActivity implements View.OnClickListener {
     public MqttCar controller;
     public CardView theoryCard, drivingCard, connectCard, supportCard;
-    //Camera Config
-    private final int IMAGE_HEIGHT = 240;
-    private final int IMAGE_WIDTH = 320;
-    public ImageView imageView;
+
 
     private BottomNavigationView bottomNavigationView;
 
@@ -41,15 +38,6 @@ public class HomeActivity extends AppCompatActivity implements View.OnClickListe
         connectCard.setOnClickListener(this);
         supportCard.setOnClickListener(this);
 
-
-        controller = new MqttCar(getApplicationContext(), () -> {
-            try {
-                controller.changeSpeed(0.5);
-            } catch (MqttException ex) {
-                ex.printStackTrace();
-            }
-        },this);
-        //imageView = findViewById(R.id.cameraView);
 
     }
 
@@ -70,27 +58,5 @@ public class HomeActivity extends AppCompatActivity implements View.OnClickListe
         break;
         default:break;
     }
-    }
-
-    /**
-     *
-     * @param message of frames to be rendered
-     * This should be called upon received a message on the Camera Topic
-     * and should then update the ImageView displayed on the current screen
-     */
-    public void cameraRendering(MqttMessage message){
-        final Bitmap bm = Bitmap.createBitmap(IMAGE_WIDTH, IMAGE_HEIGHT, Bitmap.Config.ARGB_8888);
-
-        final byte[] payload = message.getPayload();
-        final int[] colors = new int[IMAGE_WIDTH * IMAGE_HEIGHT];
-        for (int ci = 0; ci < colors.length; ++ci) {
-            final byte r = payload[3 * ci ];
-            final byte g = payload[3 * ci + 1];
-            final byte b = payload[3 * ci + 2];
-            colors[ci] = Color.rgb(r, g, b);
-        }
-
-        bm.setPixels(colors, 0, IMAGE_WIDTH, 0, 0, IMAGE_WIDTH, IMAGE_HEIGHT);
-        imageView.setImageBitmap(bm);
     }
 }
