+++ SmartCarMQTTApp/app/src/main/java/com/example/smartcarmqttapp/PracticeDrivingActivity.java
@@ -4,12 +4,32 @@ import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.Intent;
+import android.graphics.Bitmap;
+import android.graphics.Color;
 import android.os.Bundle;
 import android.view.MenuItem;
+import android.view.View;
+import android.webkit.WebSettings;
+import android.webkit.WebView;
+import android.widget.ImageView;
+import android.widget.LinearLayout;
 
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 
+import org.eclipse.paho.client.mqttv3.MqttException;
+import org.eclipse.paho.client.mqttv3.MqttMessage;
+
 public class PracticeDrivingActivity extends AppCompatActivity {
+    public MqttCar controller;
+
+    //Camera Config
+    private final int IMAGE_HEIGHT = 240;
+    private final int IMAGE_WIDTH = 320;
+    public ImageView imageView;
+    public LinearLayout webViewLayout, cameraViewLayout;
+
+    public pl.droidsonroids.gif.GifImageView screen_error;
+
 
     private BottomNavigationView bottomNavigationView;
 
@@ -51,5 +71,55 @@ public class PracticeDrivingActivity extends AppCompatActivity {
                 return false;
             }
         });
+
+
+        controller = new MqttCar(getApplicationContext(), () -> {
+            try {
+                controller.changeSpeed(0.5);
+            } catch (MqttException ex) {
+                ex.printStackTrace();
+            }
+        },this);
+
+        imageView = findViewById(R.id.cameraView);
+
+        screen_error = findViewById(R.id.screenError);
+        screen_error.setVisibility(View.GONE);
+
+
+        if(controller.isConnected()) {
+            imageView.setVisibility(View.VISIBLE);
+            screen_error.setVisibility(View.GONE);
+        }
+        else {
+            screen_error.setVisibility(View.VISIBLE);
+            imageView.setVisibility(View.GONE);
+        }
+
+
     }
+
+
+    /**
+     *
+     * @param message of frames to be rendered
+     * This should be called upon received a message on the Camera Topic
+     * and should then update the ImageView displayed on the current screen
+     */
+    public void cameraRendering(MqttMessage message){
+        final Bitmap bm = Bitmap.createBitmap(IMAGE_WIDTH, IMAGE_HEIGHT, Bitmap.Config.ARGB_8888);
+
+        final byte[] payload = message.getPayload();
+        final int[] colors = new int[IMAGE_WIDTH * IMAGE_HEIGHT];
+        for (int ci = 0; ci < colors.length; ++ci) {
+            final byte r = payload[3 * ci ];
+            final byte g = payload[3 * ci + 1];
+            final byte b = payload[3 * ci + 2];
+            colors[ci] = Color.rgb(r, g, b);
+        }
+
+        bm.setPixels(colors, 0, IMAGE_WIDTH, 0, 0, IMAGE_WIDTH, IMAGE_HEIGHT);
+        imageView.setImageBitmap(bm);
+    }
+
 }
