+++ SmartCarMQTTApp/app/src/main/res/layout/activity_practice_driving.xml
@@ -6,16 +6,36 @@
     android:layout_height="match_parent"
     tools:context=".PracticeDrivingActivity">
 
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="This is practice driving"
-        android:textSize="30dp"
-        android:layout_centerInParent="true"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="250dp"
+        android:orientation="vertical"
+        android:id="@+id/cameraViewLayout">
+
+        <ImageView
+            android:id="@+id/cameraView"
+            android:layout_width="match_parent"
+            android:layout_height="300dp"
+            android:background="@color/teamBlue" />
+
+    </LinearLayout>
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="300dp"
+        android:orientation="vertical"
+        android:id="@+id/webViewLayout">
+
+        <pl.droidsonroids.gif.GifImageView
+            android:id="@+id/screenError"
+            android:layout_width="match_parent"
+            android:layout_height="300dp"
+            android:src="@drawable/screenerror"/>
+
+
+    </LinearLayout>
+
+
 
     <com.google.android.material.bottomnavigation.BottomNavigationView
         android:id="@+id/bottom_navigation"
